Core: add --no-window-frame command line option
Add a command line option to load Blender without a window frame.

Currently this is only used on Wayland, enabled for WITH_UI_TESTS since
attempting to load LIBDECOR caused the tests to crash on start on Fedora.
For tests there is no need to use LIBDECOR, so disable the window frame.
This can also be used by users who don't want to use the X11 fallback if
LIBDECOR can't be found.

Back-ported from main so users can workaround: #144919
which won't be fixed for 5.0 release.

- 3349b97987bcaad4679667aaec6c8500038b3203
- eedf15e3b79be9ca72f133da99a97ca166ed3c6a

Ref !147716