Fix #148284, #147714: Render GPU context assert after recent fix
The change in 75f3d8c6 to use re->scene was not reliable, this is only set
in the render thread instead of the main thread immediately upon creation.
This leads to a race condition where testing for a running job is unreliable.

Instead compare with the render owner pointer that was added in the previous
refactor commit.

Pull Request: https://projects.blender.org/blender/blender/pulls/148468