Fix #146838: Windows/IME cancel composition on end IME
When clicking away from an IME input in progress it triggers back a new
`WM_IME_COMPOSITION` event which can't be handled since the button
using the IME data is already inactive. This causes a reuse of old
`WindowRuntime::ime_data` which is incorrect and fails an assert.
This PR correctly ends the IME in this case and cancels composition.

Pull Request: https://projects.blender.org/blender/blender/pulls/148072