Fix #148249: Invalid UV selection state when disabling select-sync
When disabling sync-select it was possible for single edges to be
selected in face-select mode. This happened because it was possible
for the selection data to be used even if it didn't match the
current uv select-selection mode.

Resolve this problem by converting the select-sync selection
to the non-sync selection when toggling select-sync
(generating selection data if necessary).

Also avoid unintended use of non-synchronized UV's by calling
`BM_mesh_uvselect_clear` before performing non-synchronized selection.

Ref !148560