Fix #148531: Vector Curves node 'Reset Curve' option applies offset
Introduced with 1ac1090271

The above commit incorrectly assumed that all of the curves work on X
values that range on the period [0.0, 1.0] - this is not true for the
Line and Constant Median preset types, the former of which is used as
the default for the Vector Curves node.

To fix this, simplify the behavior for the two named presets, since they
depend on the current clip region. The new implementation simply swaps
the two y values.

Additionally, asserts are added to verify the assumed state of the
preset curves.

Pull Request: https://projects.blender.org/blender/blender/pulls/148545