Fix #148436: Incorrect legacy hair offset with USD export
The report in question uncovered two sources of issues. The primary one
being that there was an accidental double-transform from placing the
basis curves prim under the main Xform of the mesh too. This is solved
by considering the inverse of mesh's world transform when writing out
the curve points.

The second was exposed with the new, more correct, viewport drawing of
curves that showed that we were exporting the wrong curve type. This
would manifest as a disappearing curve segment at the beginning and end
of the curve. Fixed by explicitly writing out catmull-rom, pinned,
curves for the hair rather than using bsplines.

As a result the Storm-USD tests now match much closer to the native
Storm-Hydra variants.

Pull Request: https://projects.blender.org/blender/blender/pulls/148543