I18n: Disambiguate "Amount" when uncountable
"Amount" in English can refer to "a number of things", or "a quantity or
volume", which is uncountable. This is not the case in all languages,
so the two senses need to be distinguished.

This commit introduces a new "Number" translation context, because the
word is not specific to any type of data, and using the existing
"Amount" context would be very confusing.

Reported by Alexandr Fatih in #43295.

Pull Request: https://projects.blender.org/blender/blender/pulls/148616