Fix #148592: EEVEE: Broken shadow unit tests on AMD
Caused by a faulty async readback.

This patch only make sure that synchronous readback
are not taking the async path. However, it doesn't
fix the core issue which is that the async path
seems to be broken.

Given the async path is only used for EEVEE shadow
statistics buffer, it is not urgent to fix.

Pull Request: https://projects.blender.org/blender/blender/pulls/148855