Fix #148822: Crash undoing with sequencer preview open
There's some code in the memfile undo system that tries to restore
dependency graphs after loading an undo step.

This apparently also restored the depsgraph built for the sequencer
preview which causes a crash later on.

This fixes the issue by discarding the depsgraph after an undo step
has been decoded.

Note that we should try to restore the VSE depsgraph (as well as the
compositor depsgraph). But this would require a bit more refactoring
of the scene `depsgraph_hash`, e.g. moving it to runtime data, etc.

Pull Request: https://projects.blender.org/blender/blender/pulls/148862