Fix: Make VSE line drawing consistent with other editors
The VSE always drew lines further apart than other editors.
This made the drawing of the scrub area vs the lines inconsistent where it felt like
the lines were just missing.
It turns out that seems to have been intentional but not documented.

In this patch I remove the modification of the grid size, but keep in mind
this is not the only possibility.
We could also *add* the same for the scrub area
drawing instead if the line density should be kept low.

Pull Request: https://projects.blender.org/blender/blender/pulls/148552