Fix #141397: Drivers created via python have a generator modifier
The issue was that when creating a driver via Python it would have a generator modifier,
meaning that there were no Keyframes that could be modified.
This is also different to how adding a driver via the UI works and was just preserved
for backwards compatibility reasons.

Introduced with bebaa410474fd79fc6f3b4ebeaea82cb170aa364

This was discussed in the Animation & Rigging module meeting:
https://devtalk.blender.org/t/2025-07-03-animation-rigging-module-meeting/41307#p-153691-ongoing-work-5

Pull Request: https://projects.blender.org/blender/blender/pulls/148875