Fix #147159: Crash selecting brush asset from current file after renaming
This sanitizes asset system code to better deal with renaming asset
data-blocks, preventing the crash. Previously, querying an asset's
library-relative path, or building an `AssetWeakReference` (which uses
the library-relative path) would get a path with the old name. Make sure
this uses the new name instead after renaming. This way the asset shelf
uses up-to-date asset weak references to be passed to the brush
activate operator.

It's not a perfect solution, since the brush won't be highlighted as
active anymore after renaming. This is because the highlighting is based
on the `AssetWeakReference` of the previously activated brush, which is
known to break on actions like renaming (hence the "weak" in the name).
It still holds the old name. I think we can make the highlighting use an
asset reference from the actual active brush to fix this. This is better
done separately, just prevent the crash for now.

Pull Request: https://projects.blender.org/blender/blender/pulls/149019