Fix: Tree View: Filter panel icon is brighter compared to UIList
This is due to `IconToggle`. Use `Toggle` so it looks consistent
with UIList. Both ButType uses different themes, so the
difference.

Julian came up with this fix for the bug that Hans reported it in chat.

See PR description for images

Pull Request: https://projects.blender.org/blender/blender/pulls/149171