Fix #149090: long warning message not readable
This just adds the same warning text as tooltip. This way, the full warning is
readable even if it does not fit in the width of the panel.

In the implementation I can't just pass the tooltip to `uiDefIconTextBut`
because that currently leads to the text being drawn twice in the tooltip. Just
creating a custom tooltip seems more straight forward compared to working around
features of automatic tooltip generation.

Pull Request: https://projects.blender.org/blender/blender/pulls/149338