Compositor: Store enabled asset shelf catalog tabs in Preferences
By default the enabled catalog tabs in the asset shelf (the visible
tabs) are saved per editor in the file. This makes sense when the assets
you deal with are quite specific to the file or project, like with poses
for example.

I'd expect that typical custom compositing assets aren't too file or
project specific, and you want to be able to access them from any
Blender session. In that case it gets annoying if you have to manage the
enabled catalog tabs for every file, even though they are usually the
same ones. Storing them in the Preferences avoids this.

Tradeoff is that in cases where assets are more project specific, the
catalogs are still shared between project files. I think in practice
that's not a big annoyance though. It's basically a "remember last"
system, so it's predictable and simple.

Also note that changes are only automatically saved with _Auto-Save
Preferences_ enabled.

Pull Request: https://projects.blender.org/blender/blender/pulls/149359