Fix #146119: UI: Some (Light) widgets lack auto-keyframing
We already had #70273 for this, and it was (partly) fixed. But only for
the for `dial`, `arrow` and `move` gizmos. This made e.g. the Camera
Lens widget work, also Light Angle was working because of that.

All others still did not auto-key (e.g. Light size [reported], Light
blend).

This is because they use `WM_gizmo_target_property_def_func` (not
`WM_gizmo_target_property_def_rna`) for which it is less transparent to
determine the actual property to key (they are not exposed as
`wmGizmoProperty->prop`). In certain cases these are multiple properties
even (Area Light size).

Resolve by introducing a new `wmGizmoPropertyFnForeachRNAProp`
which can be used to go over the RNA properties gizmos are manipulating
and has a callback (which for now does the autokeying but could be used
for other things in the future).

Used for autokeying the remaining Light gizmos in this PR (but could probably
be expanded to other things [e.g. image empties size widget, subscribing to
changes on RNA properties, tool-tips, ...]).

Also follows the general rule that after the (auto-)keying only the newly-created
keys should be selected.

Pull Request: https://projects.blender.org/blender/blender/pulls/146168