Fix #149366: Duplicating asset to current file doesn't save new catalogs
Since c22db2251e, we only write the catalogs file to disk when the
catalog service is tagged as having unsaved changes. This wasn't done
when the catalogs were created by the brush and pose asset operators.

Note that this tagging is intentionally not done in the asset system
itself to give the caller control (there are unit tests to cover this).
There are sometimes reasons to not have them tagged. Although I think we
should re-evaluate that decision, especially now that this tag is used
for more than just a UI indication.

Pull Request: https://projects.blender.org/blender/blender/pulls/149647