Fix #149599: VSE overwrite mode alters locked strips, crashes in some cases
Crash was happening when a strip was inserted in the middle of a
locked strip, with overwrite mode on. The code was calling
edit_strip_split with null error_msg parameter, but the function does
not support it being null.

Another issue is that inserting not exactly in the middle, but
overlapping at one side only was still trimming the locked strip, and
overlapping completely was deleting the locked strip. Instead, skip
altering these strips. Code later already notices that at the end of
the whole operator some strips still have "overlap" flag set on them,
and then as a fallback puts the strips being transformed onto empty
channels.

Pull Request: https://projects.blender.org/blender/blender/pulls/149628