Fix #145985: Cycles: OSL Memory leak
The shared_ptr here is effectively used to handle ownership of both
the ShadingSystem and the associated OSLRenderServices.

Therefore, it has a custom deleter to free the OSLRenderServices, but
that's the only thing it did, and since providing a deleter overrides
the default behavior, that means that the ShadingSystem itself leaked.

Pull Request: https://projects.blender.org/blender/blender/pulls/149612