Fix #147705: Slider overlapping header UI
The slider UI used to overlap the Header items in the 3D viewport.
This has been the case since 4.0 where the `RGN_TYPE_TOOL_HEADER`
was introduced and using `ED_area_status_text(pbd->area, "");` no longer
cleared the header.

The fix is to search *either* for the `RGN_TYPE_HEADER` *or* the
`RGN_TYPE_TOOL_HEADER` while preferring the latter for drawing.

As a consequence, `ED_area_status_text` has to be called by the users
of the slider UI.

In this patch I also fix the usage of `U.pixelsize` replacing it with
`UI_SCALE_FAC`.

Pull Request: https://projects.blender.org/blender/blender/pulls/149442